// src/Interfaces/ErrorLevel.ts
export type TErrorLevel = 'emergency' |
    'alert' |
    'critical' |
    'error' |
    'warning' |
    'notice' |
    'info' |
    'debug'

export enum ErrorLevel {
    EMERGENCY = 'emergency',
    ALERT = 'alert',
    CRITICAL = 'critical',
    ERROR = 'error',
    WARNING = 'warning',
    NOTICE = 'notice',
    INFO = 'info',
    DEBUG = 'debug',
    LOG = 'log',
    MESSAGE = 'message'
}